Materialize nested children when materializing parent
Nested records can not be fetched via SC.store.find unless they have
been registered with the store.  Normally this only happens when the
record is accessed.  Sometimes we need to find the record after it has
been loaded from the server (via the parent record) but before the
child has been accessed.  This ensures nested records are
registered when their parent record is materialized.